RET-5541 Fix declareXXX().value
Need to return ApptimizeValueVariable instead of base ApptimizeVariable
to have access to the .value member. The get... variants and
declareDict/Array are fine, this just looks like an oversight for the
declareSingleton variants.

SHOULD add the missing testDeclareVariableValues that would originally
have caught this; didn't because the test app is currently completely
broken - the app key no longer exists. Will be fixed separately.