RET-5542 Unpin Android sdk and switch to AAR
3.7.18+ is basically equivalent to an exact 3.7.18. Ask for
3.10+ instead. Don't automatically move to 4.0; presumably that
will be breaking in some way and will need manual upgrade.

And ask for AAR explicitly, because 3.7.18 was the last .jar release
and aar isn't automatically looked for

Also: use Apptimize.getVersion (new in 3.10.4) instead of using
Apptimize.class.getPackage().getImplementationVersion()
(which returns 0.0). Not essential, but convenient to confirm
we're now getting the expected version.